Ported search encoding to new architecture.
git-svn-id: https://svn.apache.org/repos/asf/james/server@593739 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/experimental-seda-imap-function/src/main/java/org/apache/james/experimental/imapserver/encode/writer/OutputStreamImapResponseWriter.java
@@ -77,7 +77,7 @@ public class OutputStreamImapResponseWriter extends AbstractLogEnabled implement
         }
     }
 
-    public void message( int number )
+    public void message( long number )
     {
         space();
         writer.print( number );
